+++ shared/ssz/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "decode.go",
         "encode.go",
         "encoder_decoder_lib.go",
-        "test_utils.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/shared/ssz",
     visibility = ["//visibility:public"],
