+++ shared/ssz/encode_test.go
@@ -2,7 +2,9 @@ package ssz
 
 import (
 	"bytes"
+	"encoding/hex"
 	"fmt"
+	"strings"
 	"testing"
 )
 
@@ -161,4 +163,31 @@ func TestEncodeSize(t *testing.T) {
 		size, err := EncodeSize(val)
 		return size, err
 	})
+}
+
+// unhex converts a hex string to byte array
+func unhex(str string) []byte {
+	b, err := hex.DecodeString(stripSpace(str))
+	if err != nil {
+		panic(fmt.Sprintf("invalid hex string: %q", str))
+	}
+	return b
+}
+
+func stripSpace(str string) string {
+	return strings.Replace(str, " ", "", -1)
+}
+
+type simpleStruct struct {
+	B uint16
+	A uint8
+}
+
+type innerStruct struct {
+	V uint16
+}
+
+type outerStruct struct {
+	V    uint8
+	SubV innerStruct
 }
